refactor: 💡 Refactor bnStore and add meta data
